record inner exceptions
reworked and removed the builder
resolves #15
