+++ Mindscape.Raygun4Net.Tests/Mindscape.Raygun4Net.Tests.csproj
@@ -44,7 +44,8 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="RaygunErrorMessageBuilderTests.cs" />
+    <Compile Include="RaygunErrorMessageExceptionTests.cs" />
+    <Compile Include="RaygunErrorMessageInnerExceptionTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
