--- Mindscape.Raygun4Net/IRaygunErrorMessageBuilder.cs
@@ -1,19 +1,0 @@
-﻿using System;
-
-using Mindscape.Raygun4Net.Messages;
-
-namespace Mindscape.Raygun4Net
-{
-  public interface IRaygunErrorMessageBuilder
-  {
-    RaygunErrorMessage Build();
-
-    IRaygunErrorMessageBuilder SetMessage(Exception exception);
-
-    IRaygunErrorMessageBuilder SetStackTrace(Exception exception);
-
-    IRaygunErrorMessageBuilder SetClassName(Exception exception);
-
-    IRaygunErrorMessageBuilder SetData(Exception exception);
-  }
-}
